Updated the manifest to reflect all the activities.
Updated the activity_register to reflect the same button color as the AppLauncher page.

Updated the activity_app_launcher so that the EditText text color is black (previously was white).

Updated the rounded_corner_button to reflect the same button color as the AppLauncher.

Add Intents to AppLauncherActivity, RegisterActivity, HomePageActivity.